Phase one of refactoring to enable DSM DB test containers
