+++ pepper-apis/dsm-core/src/test/java/org/broadinstitute/dsm/util/DBTestUtil.java
@@ -11,7 +11,6 @@ import java.sql.SQLException;
 import java.sql.Statement;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.UUID;
 import java.util.concurrent.ThreadLocalRandom;
 
 import com.google.gson.Gson;
