+++ pepper-apis/dsm-core/src/test/resources/testdbs/testdbsetup.cnf
@@ -1,0 +1,12 @@
+# this file is used when starting test containers
+[client]
+default-character-set=utf8
+
+[mysql]
+default-character-set=utf8
+
+
+[mysqld]
+collation-server = utf8_unicode_ci
+init-connect='SET NAMES utf8'
+character-set-server = utf8
