+++ pepper-apis/parent-pom.xml
@@ -424,7 +424,7 @@
         <dependency>
             <groupId>org.testcontainers</groupId>
             <artifactId>mysql</artifactId>
-            <version>1.12.3</version>
+            <version>1.17.6</version>
             <scope>test</scope>
         </dependency>
